v2.2.0: polish GUI, Complete/Undo clarity, dep bumps
Remove experimental Silk & Soul, Tags tab, and Wish Location scaffold.
Surface Remote Complete in Tools; refuse flag-only complete on world-state
wishes (flea carnival); Active/Done/All filters and Undo Complete for support.
Reliability: registry-driven inject via CreateEntry, wishboard TGA sweep helpers,
Gourmand/transient menu guards. Docs for Complete/Undo and Tools-only rules.

Dependencies: Silksong.FsmUtil 0.3.17, Microsoft.Unity.Analyzers 1.27.0.
Keep HarmonyX 2.9.0 and UnityEngine.Modules 6000.0.50 (game-aligned).
