+++ Directory.Build.props
@@ -1,6 +1,6 @@
 <Project>
   <PropertyGroup>
     <AssemblyTitle>QuestMod</AssemblyTitle>
-    <Version>2.1.2</Version>
+    <Version>2.2.0</Version>
   </PropertyGroup>
 </Project>
