+++ thunderstore/thunderstore.toml
@@ -4,14 +4,14 @@ schemaVersion = "0.0.1"
 [package]
 namespace = "DatUub"
 description = "Quest management mod for Silksong — accept, complete, and customize quest targets with a full debug GUI."
-version = "2.1.2"
+version = "2.2.0"
 websiteUrl = "https://github.com/DatUub/Silksong.QuestMod"
 containsNsfwContent = false
 
 [package.dependencies]
 BepInEx-BepInExPack_Silksong = "5.4.2304"
 silksong_modding-DataManager = "1.2.2"
-silksong_modding-FsmUtil = "0.3.16"
+silksong_modding-FsmUtil = "0.3.17"
 silksong_modding-UnityHelper = "1.2.0"
 
 [build]
